requirements.txt
coverage==4.5.1
Django==2.1.11
django-nose==1.4.6
djangorestframework==3.10.2
nose==1.3.7
kavenegar
django-bot-crawler-blocker
